billing: split models to files
