+++ django/src/billing/models/__init__.py
@@ -2,6 +2,9 @@ __all__ = [
     "BillingUser",
     "BillingAccount",
     "BillingTask",
+    "BillingTransaction",
 ]
 
-from .billing import *
+from .task import BillingTask
+from .transaction import BillingTransaction
+from .user import BillingUser, BillingAccount
